+++ src/main/java/recipe/service/UserService.java
@@ -1,18 +1,7 @@
 package recipe.service;
 
 import recipe.entity.User;
-import recipe.repository.UserRepository;
 
-public class UserService {
-	private final UserRepository userRepository;
-
-    public UserService(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-
-    //ID로 유저 조회 (없으면 예외 발생)
-    public User getUserById(Long userId) {
-        return userRepository.findById(userId)
-                .orElseThrow(() -> new RuntimeException("사용자를 찾을 수 없습니다."));
-    }
+public interface UserService {
+	User getUserById(Long userId);
 }
